scripts/requirements.txt
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-web==7.2.0
azure-mgmt-storage==21.2.1
azure-search-documents==11.4.0b11
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
idna==3.6
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
packaging==23.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.0
python-dotenv==1.0.1